When I click "view collection" I get a 404 error. I need a way to - view the collection contents - add links to a collection
In the future I will also want a generic "new" button that will allow me to either add a new collection, or add a link to an existing collection, or add a link to a new collection". I will provide mock-ups for all of this along with other components and flows. For now I just want to get a basic structure set-up that will allow all the necessary functionalities